- Adicionado controller PaymentType - Adicionado controller Transaction.js - Refatorado posições e adicionado novas funções globais - Adicionado mocky para serviço de autorização de cartão - Iniciado lógica de processamento da transação

+++ src/database/migrations/20200131141755-create-payment_types_table_V1.js
@@ -30,7 +30,7 @@ module.exports = {
         default: false
       },
       fee: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DOUBLE,
         default: false
       }
     });
